refactor: keep ratchet honest — split test modules, surface entity_synced
The wave-5 push tripped four readability-ratchet regressions; three get
real fixes rather than a baseline bump:
- semantics.rs (632) and sdk client.rs (524) crossed 500 lines only
  because they absorbed deduplicated code AND its tests; tests split into
  sibling _test.rs files per the repo idiom (172 and 443 lines remain).
- temper serve's mod.rs crossed 1000 after the Discord spawn grew in the
  HTTP-client consolidation; the spawn moved to serve/discord.rs (964).
- The allow(dead_code) on PersistedRecordEntity.entity_synced is gone:
  alert and insight batch items now report entity_synced per item, so a
  record whose entity mirror failed is visible in the response instead of
  only in logs.
The remaining +1 println (CLI status line in the Discord spawn, matching
that file's operator-output convention) rides an intentional baseline
snapshot; max-file-lines improved 2686 -> 2648.
