+++ .ci/readability-baseline.env
@@ -1,11 +1,11 @@
 # Generated by scripts/readability-ratchet.sh
-PROD_RS_TOTAL=485
+PROD_RS_TOTAL=486
 PROD_FILES_GT300=185
 PROD_FILES_GT500=76
 PROD_FILES_GT1000=11
-PROD_MAX_FILE_LINES=2686
+PROD_MAX_FILE_LINES=2648
 PROD_MAX_FILE_PATH=crates/temper-wasm/src/host_trait.rs
 ALLOW_CLIPPY_COUNT=30
 ALLOW_DEAD_CODE_COUNT=10
-PROD_PRINTLN_COUNT=237
+PROD_PRINTLN_COUNT=238
 PROD_UNWRAP_CI_OK_COUNT=138
